pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chess-coach"
version = "0.1.0"
description = "A conversational chess coach powered by Claude with Stockfish analysis"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-chess>=1.10.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.5.0",
]

[project.scripts]
chess-coach = "chess_coach.cli.app:cli"

[tool.setuptools.packages.find]
where = ["src"]
